Pause Menu Added
Credit page now has a back button so you aren't stuck in the credits page. I also added a pause menu. It has simple options, like to pause, resume, and to return to the main menu. A small thing to note is that the fungus talk options still appear in the pause menu.